.......... [ZBXNEXT-3386] merged from upstream
git-svn-id: svn://svn.zabbix.com/branches/dev@65077 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBXNEXT-3386/frontends/php/include/classes/api/services/CHostGeneral.php
@@ -637,9 +637,11 @@ abstract class CHostGeneral extends CHostBase {
 							' WHERE '.dbConditionInt('ia.applicationid', $applicationids).
 						')'
 				));
-				$result = API::Application()->delete(zbx_objectValues($applications, 'applicationid'), true);
-				if (!$result) {
-					self::exception(ZBX_API_ERROR_INTERNAL, _('Cannot unlink and clear applications.'));
+				if ($applications) {
+					$result = API::Application()->delete(zbx_objectValues($applications, 'applicationid'), true);
+					if (!$result) {
+						self::exception(ZBX_API_ERROR_INTERNAL, _('Cannot unlink and clear applications.'));
+					}
 				}
 			}
 			else {
